Generate render in manage-ui
