+++ agents-cli/package.json
@@ -33,7 +33,7 @@
   "license": "SEE LICENSE IN LICENSE.md",
   "dependencies": {
     "@ai-sdk/anthropic": "2.0.2",
-    "@ai-sdk/google": "^2.0.0",
+    "@ai-sdk/google": "^2.0.14",
     "@ai-sdk/openai": "2.0.11",
     "@babel/parser": "^7.23.0",
     "@babel/types": "^7.23.0",
