+++ agents-manage-ui/package.json
@@ -41,6 +41,10 @@
     "test:coverage": "vitest --run --coverage"
   },
   "dependencies": {
+    "@ai-sdk/anthropic": "2.0.2",
+    "@ai-sdk/gateway": "^1.0.23",
+    "@ai-sdk/google": "^2.0.14",
+    "@ai-sdk/openai": "2.0.11",
     "@ai-sdk/react": "2.0.11",
     "@hookform/resolvers": "^5.2.1",
     "@inkeep/agents-core": "workspace:^",
@@ -51,6 +55,7 @@
     "@nangohq/frontend": "^0.69.5",
     "@nangohq/node": "^0.69.5",
     "@nangohq/types": "^0.69.5",
+    "@openrouter/ai-sdk-provider": "^1.2.0",
     "@radix-ui/react-alert-dialog": "^1.1.15",
     "@radix-ui/react-checkbox": "^1.3.2",
     "@radix-ui/react-collapsible": "^1.1.12",
@@ -65,9 +70,11 @@
     "@radix-ui/react-switch": "^1.2.6",
     "@radix-ui/react-tabs": "^1.1.13",
     "@radix-ui/react-tooltip": "^1.2.7",
+    "@shikijs/monaco": "^3.13.0",
     "@sinclair/typebox": "^0.34.40",
     "@types/dagre": "^0.7.53",
     "@xyflow/react": "^12.8.2",
+    "ai": "^5.0.11",
     "axios": "^1.11.0",
     "axios-retry": "^4.5.0",
     "class-variance-authority": "^0.7.1",
@@ -92,9 +99,8 @@
     "react-runner": "^1.0.5",
     "recharts": "2.15.4",
     "remark-supersub": "^1.0.0",
-    "sonner": "^2.0.6",
     "shiki": "^3.13.0",
-    "@shikijs/monaco": "^3.13.0",
+    "sonner": "^2.0.6",
     "streamdown": "^1.4.0",
     "tailwind-merge": "^3.3.1",
     "use-stick-to-bottom": "^1.1.1",
